+++ packages/prettier-plugin-brace-padding/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @1adybug/prettier-plugin-block-padding
 
+## 0.0.12
+
+### Patch Changes
+
+- 新增 multipleLineBraces 选项，修复其他问题
+
 ## 0.0.11
 
 ### Patch Changes