+++ packages/prettier-plugin-remove-braces/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @1adybug/prettier-plugin-remove-braces
 
+## 0.0.6
+
+### Patch Changes
+
+- 新增 multipleLineBraces 选项，修复其他问题
+
 ## 0.0.5
 
 ### Patch Changes